[refCarInfo] Car 테이블에 refCarInfo 로 정리

+++ src/main/java/com/carshop/mycarshop/dto/car/CarInfoReqDTO.java
@@ -21,7 +21,7 @@ public class CarInfoReqDTO {
     @NotEmpty
     private String  carColors;
     @NotNull
-    private Long    carKm;
+    private int    carKm;
 
     private String mainImageFileName;
     private List<String> fileNames = new ArrayList<>();
